detectron2/requirements.txt
setuptools
Pillow
numpy
pyyaml==5.1
opencv-python
opencv-contrib-python